+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.2
+
+- Fix missing dependency.
+
 ## 0.1.1
 
 - Add support for libraries.